Refactor bus timer

+++ data/src/main/java/in/koreatech/koin/data/response/bus/BusRemainTimeResponse.kt
@@ -4,5 +4,5 @@ import com.google.gson.annotations.SerializedName
 
 data class BusRemainTimeResponse(
     @SerializedName("bus_number") val busNumber: Int? = null,
-    @SerializedName("remain_time") val remainTimeSecond: Int
+    @SerializedName("remain_time") val remainTimeSecond: Long
 )
